Move test to appropriate file and add a LinkInModule function
